Move ServicesMain to `hedera-app`
