+++ docs/docker-quickstart.md
@@ -109,7 +109,7 @@ under the `startupAccounts.path` key).
 
 As you run operations against the local network, each node will periodically save its state using
 a combination of PostgreSQL tables under _compose-network/pgdata/_ and state files under, for example,
-_compose-network/node0/saved/com.hedera.node.app.service.mono.ServicesMain/0/hedera/_.
+_compose-network/node0/saved/com.hedera.node.app.ServicesMain/0/hedera/_.
 
 To stop the network, use `Ctrl+C` (or `docker-compose stop` if running with detached containers).
