+++ hedera-node/configuration/compose/settings.txt
@@ -12,4 +12,4 @@ state.saveStatePeriod,                         300
 useLoopbackIp,                                 false
 waitAtStartup,                                 false
 virtualMap.preferredFlushQueueSize,            10000
-state.mainClassNameOverride,                   com.hedera.services.ServicesMain
+state.mainClassNameOverride,                   com.hedera.node.app.ServicesMain
