+++ hedera-node/configuration/mainnet/settings.txt
@@ -26,4 +26,4 @@ waitAtStartup,                                 false
 jasperDb.storagePath,                          /opt/hgcapp/services-hedera/HapiApp2.0/data/saved
 jasperDb.iteratorInputBufferBytes,             16777216
 virtualMap.preferredFlushQueueSize,            10000
-state.mainClassNameOverride,                   com.hedera.services.ServicesMain
+state.mainClassNameOverride,                   com.hedera.node.app.ServicesMain
