+++ hedera-node/hedera-app/build.gradle.kts
@@ -63,3 +63,91 @@ dependencies {
 tasks.withType<Test> {
     testLogging.exceptionFormat = org.gradle.api.tasks.testing.logging.TestExceptionFormat.FULL
 }
+
+// Add all the libs dependencies into the jar manifest!
+tasks.jar {
+    doFirst {
+        tasks.jar.configure {
+            manifest {
+                attributes(
+                    "Main-Class" to "com.hedera.node.app.ServicesMain",
+                    "Class-Path" to configurations.getByName("runtimeClasspath")
+                        .joinToString(separator = " ") { "../../data/lib/" + it.name }
+
+                )
+            }
+        }
+    }
+}
+
+// Replace variables in semantic-version.properties with build variables
+tasks.processResources {
+    filesMatching("semantic-version.properties") {
+        filter { line: String ->
+            if (line.contains("hapi-proto.version")) {
+                "hapi.proto.version=" + libs.versions.hapi.version.get()
+            } else if (line.contains("project.version")) {
+                "hedera.services.version=" + project.version
+            } else {
+                line
+            }
+        }
+    }
+}
+
+// Copy dependencies into `data/lib`
+val copyLib = tasks.register<Copy>("copyLib") {
+    from(project.configurations.getByName("runtimeClasspath"))
+    into(project(":hedera-node").file("data/lib"))
+}
+
+// Copy built jar into `data/apps` and rename HederaNode.jar
+val copyApp = tasks.register<Copy>("copyApp") {
+    from(tasks.jar)
+    into(project(":hedera-node").file("data/apps"))
+    rename { "HederaNode.jar" }
+    shouldRunAfter(tasks.getByName("copyLib"))
+}
+
+tasks.assemble {
+    dependsOn(copyLib)
+    dependsOn(copyApp)
+}
+
+// Create the "run" task for running a Hedera consensus node
+tasks.register<JavaExec>("run") {
+    group = "application"
+    dependsOn(tasks.assemble)
+    workingDir = project(":hedera-node").projectDir
+    jvmArgs = listOf("-cp", "data/lib/*")
+    mainClass.set("com.swirlds.platform.Browser")
+}
+
+val cleanRun = tasks.register("cleanRun") {
+    val prj = project(":hedera-node")
+    prj.delete(File(prj.projectDir, "database"))
+    prj.delete(File(prj.projectDir, "output"))
+    prj.delete(File(prj.projectDir, "settingsUsed.txt"))
+    prj.delete(File(prj.projectDir, "swirlds.jar"))
+    prj.projectDir.list { _, fileName -> fileName.startsWith("MainNetStats") }
+        ?.forEach { file ->
+            prj.delete(file)
+        }
+
+    val dataDir = File(prj.projectDir, "data")
+    prj.delete(File(dataDir, "accountBalances"))
+    prj.delete(File(dataDir, "apps"))
+    prj.delete(File(dataDir, "lib"))
+    prj.delete(File(dataDir, "recordstreams"))
+    prj.delete(File(dataDir, "saved"))
+}
+
+tasks.clean {
+    dependsOn(cleanRun)
+}
+
+tasks.register("showHapiVersion") {
+    doLast {
+        println(versionCatalogs.named("libs").findVersion("hapi-version").get().requiredVersion)
+    }
+}
