+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/state/merkle/MerkleHederaState.java
@@ -103,6 +103,11 @@ public class MerkleHederaState extends PartialNaryMerkleInternal
      */
     private final Map<String, Map<String, StateMetadata<?, ?>>> services = new HashMap<>();
 
+    // Default constructor provided for ConstructableRegistry, TO BE REMOVED ASAP
+    public MerkleHederaState() {
+
+    }
+
     /**
      * Create a new instance. This constructor must be used for all creations of this class.
      *
