+++ hedera-node/hedera-app/src/main/java/module-info.java
@@ -28,4 +28,6 @@ module com.hedera.node.app {
     requires com.hedera.node.app.service.schedule.impl;
     requires com.hedera.node.app.service.token.impl;
     requires com.hedera.node.app.service.util.impl;
+    requires com.swirlds.platform;
+    requires org.apache.logging.log4j;
 }
