+++ hedera-node/hedera-mono-service/src/main/java/com/hedera/node/app/service/mono/state/initialization/SystemAccountsCreator.java
@@ -15,17 +15,14 @@
  */
 package com.hedera.node.app.service.mono.state.initialization;
 
-import com.hedera.node.app.service.mono.ServicesMain;
 import com.hedera.node.app.service.mono.ledger.backing.BackingStore;
 import com.hedera.node.app.service.mono.state.migration.HederaAccount;
 import com.hederahashgraph.api.proto.java.AccountID;
-import com.swirlds.common.system.NodeId;
-import com.swirlds.common.system.Platform;
 import com.swirlds.common.system.address.AddressBook;
 
 public interface SystemAccountsCreator {
     /**
-     * Called in {@link ServicesMain#init(Platform, NodeId)} to ensure the network has all expected
+     * Called in ServicesMain#init(Platform, NodeId) to ensure the network has all expected
      * system accounts, especially when starting from genesis; not really a migration path, may be
      * better placed in {@code com.hedera.node.app.service.mono.ServicesState#internalInit()}.
      *
