+++ hedera-node/hedera-mono-service/src/main/java/module-info.java
@@ -1,5 +1,7 @@
 module com.hedera.node.app.service.mono {
     exports com.hedera.node.app.service.mono;
+    exports com.hedera.node.app.service.mono.grpc to
+            com.hedera.node.app;
     exports com.hedera.node.app.service.mono.state.submerkle to
             com.hedera.node.app.service.mono.testFixtures,
             com.hedera.node.app,
@@ -37,6 +39,8 @@ module com.hedera.node.app.service.mono {
             com.hedera.node.app.service.token.impl,
             com.hedera.node.app.service.token.impl.test,
             com.hedera.node.app.service.contract.impl.test;
+    exports com.hedera.node.app.service.mono.state.validation to
+            com.hedera.node.app;
     exports com.hedera.node.app.service.mono.utils.accessors;
     exports com.hedera.node.app.service.mono.sigs.utils to
             com.hedera.node.app.service.mono.testFixtures;
@@ -58,6 +62,8 @@ module com.hedera.node.app.service.mono {
     exports com.hedera.node.app.service.mono.state.enums to
             com.hedera.node.app.service.mono.testFixtures,
             com.hedera.node.app.service.token.impl.test;
+    exports com.hedera.node.app.service.mono.state.exports to
+            com.hedera.node.app;
     exports com.hedera.node.app.service.mono.records;
     exports com.hedera.node.app.service.mono.stats;
     exports com.hedera.node.app.service.mono.txns;
