+++ hedera-node/settings.txt
@@ -10,4 +10,4 @@ showInternalStats,                             1
 state.saveStatePeriod,                         300
 useLoopbackIp,                                 false
 waitAtStartup,                                 false
-state.mainClassNameOverride,                   com.hedera.services.ServicesMain
+state.mainClassNameOverride,                   com.hedera.node.app.ServicesMain
